pyproject.toml
[project]
name = "clawde"
version = "0.2.0"
description = "Audio to fingerstyle guitar tablature converter"
requires-python = ">=3.10,<3.13"
dependencies = [
    "basic-pitch[onnx]>=0.3.0",
    "pyguitarpro>=0.9",
    "mcp[cli]>=1.2.0",
    "ffmpeg-python>=0.2",
    "click>=8.0",
    "setuptools<70",
    "librosa>=0.10",
]

[project.scripts]
clawde = "clawde.cli:main"
clawde-server = "clawde.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawde"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
